% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_sample.R
\name{bvar}
\alias{bvar}
\title{Title}
\usage{
bvar(
  Yraw,
  p,
  intercept = FALSE,
  persistence = 0,
  PHI_prior = "DL",
  PHI_hyper = NULL,
  L_prior = "DL",
  L_hyper = NULL,
  draws,
  burnin,
  SV = TRUE,
  SV_hyper = NULL,
  standardize = TRUE,
  PHI_in = NULL,
  L_in = NULL
)
}
\arguments{
\item{Yraw}{data}

\item{p}{lag-length}

\item{intercept}{logical}

\item{persistence}{single value in the interval \code{[0,1]} indicating persistence
of first own-lag covariates. I.e. the prior mean for the coefficients representing
first own-lags.}

\item{PHI_prior}{character. Prior on VAR coefficients: \code{"MP"} for the Hierarchical Minnesota Prior,
\code{"SSVS"} for Stochastic Search Variable Selection or \code{"DL"} for the
Dirichlet-Laplace prior.}

\item{PHI_hyper}{list. Specify prior settings on PHI. Use \code{specify_PHI_*prior},
where \code{*} is one of {MP, SSVS} or \code{DL}.}

\item{L_prior}{character. Prior on impact matrix L (LDL' decomposition of
variance-covariance matrix): \code{"MP"} for the Hierarchical Minnesota Prior,
\code{"SSVS"} for Stochastic Search Variable Selection or \code{"DL"} for the
Dirichlet-Laplace prior.}

\item{L_hyper}{list. Specify prior settings on L. Use \code{specify_L_*prior},
where \code{*} is one of {MP, SSVS} or \code{DL}.}

\item{draws}{Integer scalar. Number of posterior draws to be saved.}

\item{burnin}{Integer scalar. Number of burnin MCMC iterations, where draws
will not be saved.}

\item{SV}{logical. If \code{TRUE} stochastic volatility specification for orthogonalized
errors, i.e. time varying variance covariance matrix \eqn{\Sigma_t=LD_t L^\prime} is assumed.
If \code{FALSE} constant variance-covariance matrix is assumed with
inverse Gamma prior with \code{scale=0.01} and \code{rate=0.01}
on diagonal elements of D.}

\item{SV_hyper}{list. Use \code{specify_priors}; imported helper function from
package \code{stochvol}.}

\item{standardize}{logical. If \code{TRUE}, then data is standardized to have
zero mean and unit variance. That is the default, because priors are not invariant
to different scales of the data.}
}
\value{
Something
}
\description{
Title
}
