% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Data from the US-economy}
\format{
A 'xts' object with 247 rows and 21 variables.
}
\source{
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
dat
}
\description{
21 selected quarterly time-series from 1953:Q1 to 2021:Q2. From FRED-QD data
base. Release date 2021-07. Data is transformed to be approximately
stationary (using the transformations proposed in
\url{https://doi.org/10.20955/r.103.1-44}.
}
\keyword{datasets}
