% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_growth}
\alias{dat_growth}
\title{Data from the US-economy}
\format{
A 'xts' object with 247 rows and 21 variables.
}
\source{
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
dat_growth
}
\description{
21 selected quarterly time-series from 1953:Q1 to 2021:Q2. From FRED-QD data
base. Release date 2021-07. Data is transformed to be interpreted as
growth-rates (first log-differences with the exception of interest rates,
which are already growth rates).
}
\keyword{datasets}
