% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_raw}
\alias{data_raw}
\title{Data from the US-economy}
\format{
A 'xts' object with 250 rows and 21 variables.
}
\source{
\url{https://research.stlouisfed.org/econ/mccracken/fred-databases/}
}
\usage{
data_raw
}
\description{
21 selected quarterly time-series from 1953:Q1 to 2021:Q2. From FRED-QD data
base. Release date 2021-07.
}
\keyword{datasets}
