% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{extract_B0}
\alias{extract_B0}
\title{Retrieve the structural parameter \eqn{\boldsymbol{B}_0} samples from an IRF object.}
\usage{
extract_B0(x)
}
\arguments{
\item{x}{a \code{bayesianVARs_irf} object}
}
\description{
Retrieve the structural parameter \eqn{\boldsymbol{B}_0} samples from an IRF object.
}
\seealso{
\code{\link{specify_structural_restrictions}}
}
\author{
Stefan Haan \email{sthaan@edu.aau.at}
}
