% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse response functions}
\usage{
irf(
  x,
  ahead = 8,
  structural_restrictions = NULL,
  shocks = NULL,
  hairy = FALSE,
  t = nrow(x$logvar),
  ...
)
}
\arguments{
\item{x}{An object of type \code{\link{bayesianVARs_bvar}}.}

\item{ahead}{maximum number of time horizons.}

\item{structural_restrictions}{an object generated by \code{\link{specify_structural_restrictions}}. If specified,
the IRFs to the structural shocks identified by the restrictions given in this argument will be calculated.
If not specified, the orthogonal IRFs based on the Cholesky decomposition will be calculated.
Note that the orthogonal IRFs depend on the ordering of the variables and do not necessarily
correspond to shocks with a meaningful interpretation.}

\item{shocks}{an matrix with M rows, where M is the VAR dimension. Each column specifies a shock.
Default: \code{diag(M)}, will calculate the IRFs to M structural shocks of one standard deviation.}

\item{hairy}{set to \code{TRUE} in order to plot each path seperately.
To show valid quantiles, an Bayes optimal order of the posterior samples will be calculated which can take a long time
even for moderately many samples. Default: \code{FALSE}.}
}
\value{
Returns a \code{bayesianVARs_irf} object.
}
\description{
Effect of the structural (factor) shocks over time.
}
\details{
If a factor model was used, then the number of shocks is equal to the number
of factors.

If the Cholesky model was used, then the number of shocks is equal to the
number of variables.
}
\examples{
```{r}
train_data <- 100 * usmacro_growth[,c("GDPC1", "GDPCTPI", "GS1", "M2REAL", "CPIAUCSL")]
prior_sigma <- specify_prior_sigma(M=ncol(train_data), type="cholesky", cholesky_heteroscedastic=FALSE)
mod <- bvar(train_data, lags=5L, draws=3000, prior_sigma=prior_sigma)

structural_restrictions <- specify_structural_restrictions(
 mod,
 restrictions_B0=rbind(
   c(1 ,NA,0 ,NA,NA),
   c(0 ,1 ,0 ,NA,NA),
   c(0 ,NA,1 ,NA,NA),
   c(0 ,0 ,NA,1 ,NA),
   c(0 ,0 ,0 ,0 ,1 )
 )
)
irf_structural <- irf(
 mod, ahead=8,
 structural_restrictions=structural_restrictions
)
plot(irf_structural)
```

}
\seealso{
\code{\link{specify_structural_restrictions}}
}
\author{
Stefan Haan \email{sthaan@edu.aau.at}
}
