% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.L}
\alias{plot.L}
\title{Plot method for contemporaneous coefficients}
\usage{
\method{plot}{L}(
  object,
  summary = "median",
  ylabels = NULL,
  xlabels = NULL,
  add_numbers = TRUE,
  zlim = NULL,
  main = "",
  ...
)
}
\arguments{
\item{object}{\code{L}-object obtained from \code{\link{bvar}}.}

\item{summary}{character indicating the posterior summary to be visualized.
One of \code{"median"}, \code{"mean"}, \code{"IQR"}, \code{"sd"} or
\code{"var"}.}

\item{ylabels}{\code{NULL}, the default, indicates that ylabels will be
displayed. \code{""} indicates that no ylabels will be displayed.}

\item{xlabels}{\code{NULL}, the default, indicates that xlabels will be
displayed. \code{""} indicates that no xlabels will be displayed.}

\item{add_numbers}{logical. \code{add_numbers=TRUE}, the default indicates
that the actual values of \code{summary} will be displayed.}

\item{zlim}{numeric vector of length two indicating the minimum and maximum
values for which colors should be plotted. By default this range is
determined by the maximum of the absolute values of the selected summary.}

\item{...}{}
}
\description{
Plot method for contemporaneous coefficients
}
