% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.summary.bayesianVARs_bvar}
\alias{print.summary.bayesianVARs_bvar}
\title{Print method for summary.bayesianVARs_bvar objects}
\usage{
\method{print}{summary.bayesianVARs_bvar}(x, ...)
}
\arguments{
\item{x}{A \code{summary.bayesianVARs_bvar} object obtained via
\code{\link[=summary.bayesianVARs_bvar]{summary.bayesianVARs_bvar()}}.}

\item{...}{Currently ignored!}
}
\value{
Returns invisibly!
}
\description{
Print method for \code{summary.bayesianVARs_bvar} objects.
}
\examples{
# Access a subset of the usmacro_growth dataset
data <- usmacro_growth[,c("GDPC1", "CPIAUCSL", "FEDFUNDS")]

# Estimate model
mod <- bvar(data)

# Print summary
summary(mod)
}
