% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.summary.bayesianVARs_predict}
\alias{print.summary.bayesianVARs_predict}
\title{Print method for summary.bayesianVARs_predict objects}
\usage{
\method{print}{summary.bayesianVARs_predict}(x, ...)
}
\arguments{
\item{x}{A \code{summary.bayesianVARs_predict} object obtained via \code{\link[=summary.bayesianVARs_predict]{summary.bayesianVARs_predict()}}.}

\item{...}{Currently ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for \code{summary.bayesianVARs_predict} objects.
}
\examples{
# Access a subset of the usmacro_growth dataset
data <- usmacro_growth[,c("GDPC1", "CPIAUCSL", "FEDFUNDS")]

# Split data in train and test
train <- data[1:(nrow(data)-4),]
test <- data[-c(1:(nrow(data)-4)),]

# Estimate model using train data only
mod <- bvar(train)

# Simulate from 1-step ahead posterior predictive
predictions <- predict(mod, ahead = 1L)
sum <- summary(predictions)
print(sum)

}
