% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_PHI_DLprior}
\alias{specify_PHI_DLprior}
\title{Specify hyperparameters for Dirichlet-Laplace prior on VAR coefficients}
\usage{
specify_PHI_DLprior(a = 0.5, hyperhyper = FALSE)
}
\arguments{
\item{a}{single non-negative number, indicating the concentration parameter of
the DL prior. Only necessary if \code{hyperhyper=FALSE}. Good properties somewhere in
the interval \code{[1/n,1/2]}, where \code{n} is the number of VAR coefficients.
Smaller values imply heavier regularization towards zero.}

\item{hyperhyper}{logical. \code{TRUE} imposes a discrete uniform hyperprior on
the concentration parameter on the interval \code{[1/n,1/2]} with 1000 support points.
If set to \code{FALSE}, \code{a} has to be specified.}
}
\value{
list
}
\description{
Specify hyperparameters for Dirichlet-Laplace prior on VAR coefficients
}
