% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_priorL}
\alias{specify_priorL}
\title{Specify prior on L}
\usage{
specify_priorL(
  prior,
  DL_b = "1/n",
  R2D2_b = 0.5,
  SSVS_c0 = 0.001,
  SSVS_c1 = 1,
  SSVS_sa = 0.5,
  SSVS_sb = 0.5,
  HMP_lambda3 = c(0.01, 0.01),
  V_i = 10,
  ...
)
}
\arguments{
\item{prior}{}

\item{...}{}
}
\description{
Specify
}
