% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_priorPHI}
\alias{specify_priorPHI}
\title{Specify prior on PHI}
\usage{
specify_priorPHI(
  prior,
  DL_a = "1/K",
  R2D2_b = 0.5,
  SSVS_c0 = 0.01,
  SSVS_c1 = 100,
  SSVS_semiautomatic = TRUE,
  SSVS_sa = 0.5,
  SSVS_sb = 0.5,
  HMP_lambda1 = c(0.01, 0.01),
  HMP_lambda2 = c(0.01, 0.01),
  V_i = 10,
  global_local_grouping = "global",
  ...
)
}
\arguments{
\item{prior}{}

\item{...}{}
}
\description{
Specify
}
