% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{specify_structural_restrictions}
\alias{specify_structural_restrictions}
\title{Identifying restrictions for the structural parameters}
\usage{
specify_structural_restrictions(
  x,
  restrictions_facload = NULL,
  restrictions_B0_inv_t = NULL,
  restrictions_B0 = NULL,
  restrictions_structural_coeff = NULL,
  restrictions_long_run_ir = NULL
)
}
\arguments{
\item{restrictions_*}{A matrix. The entries have following meaning
\describe{
\item{NA:}{an unrestricted entry.}
\item{0:}{The entry at this position is restricted to be zero (i.e. an exclusion restriction).}
\item{A positive number:}{The sign of this entry should be positive.}
\item{A negative number:}{The sign of this entry should be negative.}
}}
}
\description{
Identifying restrictions for the structural parameters
}
