% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_wrapper.R
\name{[.bayesianVARs_coef}
\alias{[.bayesianVARs_coef}
\title{Extract or Replace Parts of a bayesianVARs_coef object}
\usage{
\method{[}{bayesianVARs_coef}(x, i, j, ...)
}
\arguments{
\item{x}{An object of type \code{bayesianVARs_coef}.}

\item{i}{indices}

\item{j}{indices}

\item{...}{further indices}
}
\value{
An object of type \code{bayesianVARs_coef}.
}
\description{
Extract or replace parts of a \code{bayesianVARs_coef} object.
}
\examples{
# Access a subset of the usmacro_growth dataset
data <- usmacro_growth[,c("GDPC1", "CPIAUCSL", "FEDFUNDS")]

# Estimate a model
mod <- bvar(data, sv_keep = "all")

# Extract coefficients, which are of class bayesianVARs_coef
phi <- coef(mod)
phi[1,1,1]
}
