% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_wrapper.R
\name{bvar}
\alias{bvar}
\title{Markov Chain Monte Carlo Sampling for Bayesian Vectorautoregressions}
\usage{
bvar(
  Yraw,
  p,
  intercept = 100,
  draws = 1000,
  burnin = 1000,
  persistence = 0,
  priorPHI,
  priorL,
  SV = TRUE,
  sv_spec = list(priormu = c(0, 100), priorphi = c(20, 1.5), priorsigma2 = c(0.5, 0.5),
    sv_offset = 0),
  priorHomoscedastic = NA,
  progressbar = TRUE,
  PHI_in = NULL,
  L_in = NULL
)
}
\arguments{
\item{Yraw}{Data matrix (can be a time series object). Each of \eqn{M} columns
is assumed to contain a single time-series of length \eqn{T}.}

\item{p}{Integer indicating the order of the VAR, i.e. the number of lags of
the dependent variables included as predictors.}

\item{intercept}{Either \code{intercept=FALSE} and no constant term
(intercept) will be included. Or a numeric vector of length \eqn{M}
indicating the (fixed) prior variances on the constant term. A single number
will be recycled accordingly. Default is \code{intercept=100}, which imposes
hardly any regularization w.r.t. the intercept.}

\item{draws}{single number indicating the number of draws after the burnin}

\item{burnin}{single number indicating the number of draws discarded as burnin}

\item{persistence}{single number greater or equal to 0 indicating the prior
mean of the first own-lag coefficients (default is 0).}

\item{priorPHI}{List object from \code{\link{specify_priorPHI}} with prior
settings. Used to select and customize the prior on the reduced-form VAR
coefficients.}

\item{priorL}{List object from \code{\link{specify_priorL}} with prior
settings. Used to select and customize the prior on the constant covariance
(precision) parameters.}

\item{SV}{logical indicating whether time-varying (\code{TRUE}) or constant
(\code{FALSE}) variance should be estimated.}

\item{sv_spec}{list with stochastic volatility specification if
\code{SV=TRUE}. Must contain}

\item{priorHomoscedastic}{Only used if \code{SV=FALSE}. In that case
\code{priorHomoscedastic} must be a \eqn{M \times 2} matrix, where the
first column entries indicate the shape and the second column entries
indicate the scale parameters of the inverse Gamma prior distributions of
the orthogonalized variances. All entries must be greater than 0. A vector
of length 2 will be recycled accordingly.}

\item{progressbar}{logical value indicating whether a progressbar should be
displayed during sampling (default is \code{TRUE}).}

\item{PHI_in}{numeric matrix used as initial draw for the reduced-form VAR
coefficients. Must be of dimension c(K,M), where K is the number of
covariables per equation and M the number of endogenous variables. If not
specified the posterior mean of a flat conjugate prior will be used.}

\item{L_in}{numeric upper unitriangular matrix (with ones on the diagonal) of
dimension c(M,M). Used as initial draw for covariance paramters. If not
specified the normalized cholesky factor of the posterior mean of the
variance-covariance matrix of a flat conjugate prior will be used.}
}
\description{
\code{bvar} simulates from the joint posterior distribution of the parameters
and latent variables and returns the posterior draws.
}
\details{
The VAR(p) model is of the following form:  \eqn{ \bold{y}^\prime_t =
\bold{x}^\prime_t\bold{\Phi} + \bold{\epsilon}^\prime_t}, where
\eqn{\bold{y}_t} is a \eqn{M \times 1} vector of dependent variables and
\eqn{\bold{\epsilon}_t} is the error term of the same dimension.
\eqn{\bold{x}_t} is a \eqn{K \times 1} vector containing lagged/past values of
the independent variables \eqn{\bold{y}_{t-l}} for \eqn{l=1,\dots,p} and if
specified the constant (intercept) term. If the constant term is included
\eqn{K=pM+1}, else \eqn{K=pM}. The reduced-form coefficient matrix
\eqn{\bold{\Phi}} is of dimension \eqn{K \times M}.

The disturbances \eqn{\bold{\epsilon}_t} are assumed to follow a
\eqn{M}-dimensional multivariate normal distribution with zero mean and
variance-covariance matrix \eqn{\bold{\Sigma}_t}. This matrix can be
decomposed as \eqn{\bold{\Sigma}_t = \bold{L}^{\prime -1} \bold{D}_t
\bold{L}^{-1}}, where \eqn{\bold{L}^{-1}} is upper unitriangular (with ones on
the diagonal) and \eqn{\bold{D}_t} diagonal. If \code{SV=TRUE} the matrix
\eqn{\bold{D}_t} is assumed to be time-varying and the orthogonalized
disturbances \eqn{\bold{\xi}^\prime_t = \bold{\epsilon}_t^\prime\bold{L}} will
be modeled with stochastic volatility specification. By setting
\code{SV=FALSE} a constant matrix \eqn{\bold{D}_t=\bold{D}} will be estimated.
}
\section{MCMC algorithm}{
 To sample efficiently the reduced form VAR
coefficients, the corrected triangular algorithm in Carriero et al. (2021)
is implemented. The SV parameters and latent variables are sampled using
package \code{\link{stochvol}}'s \code{\link[stochvol]{update_fast_sv}}
function. The precision parameters, i.e. the free off-diagonal elements in
\eqn{\bold{L}}, are sampled as in Cogley and Sargent (2005).
}

\references{
Carriero, A. and Chan, J. and  Clark, T. E. and Marcellino, M.
(2021). Corrigendum to “Large Bayesian vector autoregressions with
stochastic volatility and non-conjugate priors” [J. Econometrics 212 (1)
(2019) 137–154]. \emph{Journal of Econometrics},
\doi{10.1016/j.jeconom.2021.11.010}.

Cogley, S. and Sargent, T. (2005). Drifts and volatilities:
monetary policies and outcomes in the post WWII US. \emph{Review of Economic
Dynamics}, \bold{8}, 262--302, \doi{10.1016/j.red.2004.10.009}.

Hosszejni, D. and Kastner, G. (2021). Modeling Univariate and
Multivariate Stochastic Volatility in R with stochvol and factorstochvol.
\emph{Journal of Statistical Software}, \emph{100}, 1–-34.
\doi{10.18637/jss.v100.i12}.
}
