% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{find_rotation}
\alias{find_rotation}
\title{Identifying restrictions for the structural parameters}
\usage{
find_rotation(
  x,
  restrictions_facload = NULL,
  restrictions_B0_inv_t = NULL,
  restrictions_B0 = NULL,
  restrictions_structural_coeff = NULL,
  restrictions_long_run_ir = NULL,
  solver = "randomized",
  randomized_max_attempts = 100
)
}
\arguments{
\item{restrictions_*}{A matrix. The entries have following meaning
\describe{
\item{NA:}{an unrestricted entry.}
\item{0:}{The entry at this position is restricted to be zero (i.e. an exclusion restriction).}
\item{A positive number:}{The sign of this entry should be positive.}
\item{A negative number:}{The sign of this entry should be negative.}
}}
}
\description{
Identifying restrictions for the structural parameters
}
\examples{
train_data <- 100 * usmacro_growth[,c("GDPC1", "GPDICTPI", "GS1", "M2REAL", "CPIAUCSL")]
prior_sigma <- specify_prior_sigma(data=train_data, type="cholesky")
x <- bvar(train_data, lags=2L, draws=10000, prior_sigma=prior_sigma)

restrictions_B0 <- cbind(
	c(1, 0, 0, 0, 0),
	c(NA, 1, 0, 0, 0),
	c(NA, 0, NA, NA, NA),
	c(NA, NA, NA, NA, NA),
	c(NA, NA, NA, NA, NA)
)

restrictions_IR_inf <- cbind(
	c(NA, NA, NA, NA, NA),
	c(NA, NA, NA, NA, NA),
	c(0, NA, NA, NA, NA),
	c(NA, NA, NA, NA, NA),
	c(NA, NA, NA, NA, NA)
)

rotation <- find_rotation(x,
	restrictions_B0 = restrictions_B0,
	restrictions_long_run_ir = restrictions_IR_inf
)
plot(irf(x, rotation=rotation))
}
