% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse response functions}
\usage{
irf(x, factor_shock, ahead = 8)
}
\arguments{
\item{x}{An object of type \code{bayesianVARs_bvar}}
}
\description{
Effect of a shock on the variables over time
}
