% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse response functions}
\usage{
irf(x, ahead = 8, rotation = NULL, shocks = NULL)
}
\arguments{
\item{x}{An object of type \code{bayesianVARs_bvar}.}
}
\value{
Returns a \code{bayesianVARs_irf} object
}
\description{
Effect of a shock on the factors or variables over time
}
\details{
If a factor model was used, then the number of shocks is equal to the number
of factors.

If the Cholesky model was used, then the number of shocks is equal to the
number of variables / time series.
}
\examples{
train_data <- 100 * usmacro_growth[,c("GDPC1", "PCECC96", "GPDIC1", "AWHMAN", "GDPCTPI", "CES2000000008x", "FEDFUNDS", "GS10", "EXUSUKx", "S&P 500")]
prior_sigma <- specify_prior_sigma(data=train_data, type="cholesky")
mod <- bvar(train_data, lags=2L, draws=2000, prior_sigma=prior_sigma) 
plot(irf(mod))

}
