% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_gig}
\alias{my_gig}
\title{Draw from generalized inverse Gaussian}
\usage{
my_gig(n, lambda, chi, psi)
}
\arguments{
\item{n}{A single integer indicating the number of draws to generate.}

\item{lambda}{vector of shape parameters.}

\item{chi}{vector of shape/scale parameters. Must be nonnegative for positive lambdas and positive else.}

\item{psi}{vector of shape/scale parameters. Must be nonnegative for negative lambdas and positive else.}
}
\description{
Vectorized version of \code{\link[GIGrvg]{rgig}}
}
