% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{predict.bvar}
\alias{predict.bvar}
\title{Predict method for Bayesian VARs}
\usage{
\method{predict}{bvar}(object, nsteps, LPL = FALSE, Y_obs = NA, LPL_VoI = NA, ...)
}
\arguments{
\item{object}{A \code{bvar} object, obtained from \code{\link{bvar}}.}

\item{nsteps}{single positive integer indicating the forecasting horizon.}

\item{LPL}{logical indicating whether log predictive likelihood should be
computed. If \code{LPL=TRUE}, \code{Y_obs} has to be specified.}

\item{Y_obs}{Data matrix of observed values for computation of LPL. Each of
\eqn{M} columns is assumed to contain a single time-series of length
\code{nsteps}.}

\item{LPL_VoI}{either integer vector or character vector of column-names
indicating for which subgroup of time-series in \code{Yraw} from
\code{bvar}-object a joint LPL will be returned.}

\item{...}{Do not use!}
}
\description{
Simulates from (out-of-sample) predictive density for Bayesian VARs estimated
via \code{\link{bvar}} and computes log predictive likelhoods if ex-post
observed data is supplied.
}
