% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_PHI_DLprior}
\alias{specify_PHI_DLprior}
\title{Specify hyperparameters for Dirichlet-Laplace prior on VAR coefficients}
\usage{
specify_PHI_DLprior(a = 0.5, hyperhyper = FALSE)
}
\arguments{
\item{a}{numeric constant. Concentration parameter, that dictates the shrinkage. Only necessary if ´hyperhyper = FALSE'.}

\item{hyperhyper}{bool. TRUE: Discrete uniform hyperprior on a. FALSE: `a' has to be defined.}
}
\value{
list
}
\description{
Specify hyperparameters for Dirichlet-Laplace prior on VAR coefficients
}
