% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_priorL}
\alias{specify_priorL}
\title{Specify prior on L}
\usage{
specify_priorL(
  prior,
  DL_b = "1/n",
  R2D2_b = 0.5,
  SSVS_c0 = 0.001,
  SSVS_c1 = 1,
  SSVS_p = 0.5,
  HMP_lambda3 = c(0.01, 0.01),
  V_i = 10,
  ...
)
}
\arguments{
\item{prior}{character, one of \code{"R2D2"}, \code{"DL"}, \code{"SSVS"},
\code{"HMP"} or \code{"normal"}.}

\item{DL_b}{either single positive number -- where smaller values indicate
heavier regularization --, or one of \code{DL_b="1/n"} or
\code{DL_b="hyperprior"}. n is the number of free off-diagonal elements in
L.  In case of \code{DL_b="hyperprior"} a discrete uniform hyperprior is
placed on the parameter. \code{DL_b} has only to be specified if
\code{prior="DL"}.}

\item{R2D2_b}{either single positive number -- where greater values indicate
heavier regularization --, or \code{R2D2_b="hyperprior"}. In case of the
latter a discrete uniform hyperprior is placed on the parameter.
\code{R2D2_b} has only to be specified if \code{prior="R2D2"}.}

\item{SSVS_c0}{single positive number indicating the standard deviation of
the spike component. \code{SSVS_c0} has only to be specified if
\code{prior="SSVS"}. It should be that \eqn{SSVS_{c0} \ll SSVS_{c1}}!}

\item{SSVS_c1}{single positive number indicating the standard deviation of
the slab component. \code{SSVS_c1} has only to be specified if
\code{prior="SSVS"}. It should be that \eqn{SSVS_{c0} \ll SSVS_{c1}}!}

\item{HMP_lambda3}{numeric vector of length 2. Both entries must be positive.
The first indicates the shape and the second the rate of the Gamma
hyperprior on the free off-diagonal elements in L. \code{HMP_lambda3} has
only to be specified if \code{prior="HMP"}.}

\item{V_i}{numeric vector of length n, where n is the number of free
off-diagonal elements in L, indicating the prior variances. A single number
will be recycled accordingly! Must be positive. \code{V_i} has only to be
specified if \code{prior="HMP"}.}

\item{...}{Do not use!}

\item{SSVS_sa}{positive number in the range \eqn{(0,1)} indicating the first
shape parameter of the Beta hyperprior on the prior-inclusion
probability.\code{SSVS_sa} has only to be specified if \code{prior="SSVS"}.}

\item{SSVS_sb}{positive number in the range \eqn{(0,1)} indicating the second
shape parameter of the Beta hyperprior on the prior-inclusion
probability.\code{SSVS_sb} has only to be specified if \code{prior="SSVS"}.}
}
\description{
Configures prior on L.
}
